requirements.txt
web3>=6.0.0
eth-account>=0.9.0
colorama>=0.4.6
requests>=2.28.0
asyncio>=3.4.3
aiodns>=1.2.0
aiohttp[speedups]>=3.8.0
loguru>=0.7.0
playwright>=1.40.0
python-dotenv>=1.0.0
PySocks>=1.7.1

